Check for and resolve duplicate DOM IDs
Added a check to see if CSS had already been set for the default ID and if it has, starts incrementing from 1 trying to find an ID without a match. Currently taps out at 99 - this seems like a reasonable limit since the incrementer will only get invoked if the same map is used on the same page multiple times. 99 of the same map on the same page seems like a reasonable limit to me.